Exercises 4.15 - 4.21 done

+++ osa4/blogilista/controllers/blogs.js
@@ -1,24 +1,45 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const select = { username: 1, name: 1 }
+  const blogs = await Blog.find({}).populate('user', select)
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+  const body = request.body
 
-  if (!blog.hasOwnProperty('likes')) {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }  
+  const user = await User.findById(decodedToken.id)
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
+
+  if (!blog.title) {
+    return response.status(400).json({ error: 'title missing' })
+  } else if (!blog.url) {
+    return response.status(400).json({ error: 'url missing' })
+  }
+
+  if (!blog.likes) {
     blog.likes = 0
   }
 
-  if (!blog.title || !blog.url) {
-    response.status(400).end()
-  } else {
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
-  }
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -29,14 +50,24 @@ blogsRouter.put('/:id', async (request, response) => {
     likes: request.body.likes 
   }
   const options = { new: true }
-  
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, update, options)
   response.json(updatedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }  
+
+  const user = (await User.findById(decodedToken.id)).toJSON()
+  const blog = (await Blog.findById(request.params.id)).toJSON()
+
+  if (user.id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  return response.status(401).json({ error: 'blog does not belong to user' })
 })
 
 module.exports = blogsRouter
